Remove unnecessary casts from int to long

+++ src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/plugins/BMPReadWriterPlugin.java
@@ -23,7 +23,7 @@ public final class BMPReadWriterPlugin extends AbstractPrimitiveFactoryHolder {
     protected abstract static class AbstractBMPPluginNode extends AbstractPrimitiveNode {
         protected static final boolean inBounds(final long formBitsIndex, final long width, final NativeObject formBits, final NativeObject pixelLine) {
             return 1 <= formBitsIndex && formBitsIndex <= formBits.getIntLength() && 0 <= width &&
-                            width <= (long) formBits.getIntLength() - formBitsIndex + 1 && width <= pixelLine.getByteLength() / 3;
+                            width <= formBits.getIntLength() - formBitsIndex + 1 && width <= pixelLine.getByteLength() / 3;
         }
     }
